# api-docs/api/swagger/swagger.yaml
openapi: 3.0.0
info:
    title: 'Guardian'
    description: 'The Guardian is a modular open-source solution that includes best-in-class identity management and decentralized ledger technology (DLT) libraries. At the heart of the Guardian solution is a sophisticated Policy Workflow Engine (PWE) that enables applications to offer a requirements-based tokenization implementation.'
    version: '1.2.1'
    contact:
        name: 'API developer'
        url: 'https://envisionblockchain.com'
        email: 'info@envisionblockchain.com'
    license:
        name: 'Apache 2.0'
        url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
    - url: /api/v1
      description: version 1.0
components:
    securitySchemes:
        bearerAuth:
            type: 'http'
            scheme: 'bearer'
            bearerFormat: 'JWT'
    schemas:
...
        Error:
            type: object
            required:
                - code
                - message
            properties:
                code:
                    type: number
                message:
                    type: string
...
        CommonSettings:
            type: object
            properties:
                operatorId:
                    type: string
                operatorKey:
                    type: string
                schemaTopicId:
                    type: string
                nftApiKey:
                    type: string
...
paths:
...
    /settings:
        get:
            tags:
                - settings
            description: Returns current settings. For users with the Root Authority role only.
            security:
                - bearerAuth: []
            summary: Returns current settings.
            responses:
                200:
                    description: Successful operation.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CommonSettings'
                401:
                    description: Unauthorized.
                403:
                    description: Forbidden.
                500:
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        post:
            tags:
                - settings
            description: Set settings. For users with the Root Authority role only.
            security:
                - bearerAuth: []
            requestBody:
                description: Settings.
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CommonSettings'
            summary: Set settings.
            responses:
                200:
                    description: Successful operation.
                401:
                    description: Unauthorized.
                403:
                    description: Forbidden.
                500:
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...